+++ src/physics/PointMass.h
@@ -9,12 +9,12 @@
 namespace Physics {
     class PointMass : public PhysicsBody{
     public:
-        explicit PointMass(uint32_t id, float m, glm::vec3 pos = glm::vec3(0.0f), bool isStatic = false);
+        explicit PointMass(uint32_t id, double m, glm::vec3 pos = glm::vec3(0.0f), bool isStatic = false);
         explicit PointMass(uint32_t id, glm::vec3 pos = glm::vec3(0.0f), bool isStatic = true); // static objects dont need mass
 
         void applyImpulse(const glm::vec3& impulse, BodyLock lock);
 
-        void setMass(float newMass, BodyLock lock) override;
+        void setMass(double newMass, BodyLock lock) override;
         void setThermalProperty(const ThermalProperties& newProps, BodyLock lock) override;
 
         void step(float dt, BodyLock lock) override;
